Several changes.

+++ src/cpp/engine/camera.hpp
@@ -43,32 +43,32 @@ class CameraTransform : public Transform {
 /// The base class for all cameras
 class Camera : public GameObject {
  public:
-  Camera(GameObject* parent, double fovy, double z_near, double z_far);
+  Camera(GameObject* parent, float fovy, float z_near, float z_far);
   virtual ~Camera() {}
 
   virtual void ScreenResized(size_t width, size_t height) override;
 
-  const glm::dmat4& cameraMatrix() const { return cam_mat_; }
-  const glm::dmat4& projectionMatrix() const { return proj_mat_; }
+  const glm::mat4& cameraMatrix() const { return cam_mat_; }
+  const glm::mat4& projectionMatrix() const { return proj_mat_; }
   const Frustum& frustum() const { return frustum_; }
 
-  double fovx() const { return fovy_*width_/height_;}
-  void set_fovx(double fovx) { fovy_ = fovx*height_/width_; }
-  double fovy() const { return fovy_;}
-  void set_fovy(double fovy) { fovy_ = fovy; }
-  double z_near() const { return z_near_;}
-  void set_z_near(double z_near) { z_near_ = z_near; }
-  double z_far() const { return z_far_;}
-  void set_z_far(double z_far) { z_far_ = z_far; }
+  float fovx() const { return fovy_*width_/height_;}
+  void set_fovx(float fovx) { fovy_ = fovx*height_/width_; }
+  float fovy() const { return fovy_;}
+  void set_fovy(float fovy) { fovy_ = fovy; }
+  float z_near() const { return z_near_;}
+  void set_z_near(float z_near) { z_near_ = z_near; }
+  float z_far() const { return z_far_;}
+  void set_z_far(float z_far) { z_far_ = z_far; }
 
  protected:
   // it must be called through Update()
   void UpdateCache();
 
  private:
-  double fovy_, z_near_, z_far_, width_, height_;
+  float fovy_, z_near_, z_far_, width_, height_;
 
-  glm::dmat4 cam_mat_, proj_mat_;
+  glm::mat4 cam_mat_, proj_mat_;
   Frustum frustum_;
 
   void UpdateCameraMatrix();
@@ -78,23 +78,23 @@ class Camera : public GameObject {
 
 class FreeFlyCamera : public Camera {
  public:
-  FreeFlyCamera(GameObject* parent, double fov, double z_near,
-                double z_far, const glm::dvec3& pos,
-                const glm::dvec3& target = glm::dvec3(),
-                double speed_per_sec = 5.0f,
-                double mouse_sensitivity = 5.0f);
+  FreeFlyCamera(GameObject* parent, float fov, float z_near,
+                float z_far, const glm::vec3& pos,
+                const glm::vec3& target = glm::vec3(),
+                float speed_per_sec = 5.0f,
+                float mouse_sensitivity = 5.0f);
 
-  double speed_per_sec() const { return speed_per_sec_; }
-  double mouse_sensitivity() const { return mouse_sensitivity_; }
-  double cos_max_pitch_angle() const { return cos_max_pitch_angle_; }
+  float speed_per_sec() const { return speed_per_sec_; }
+  float mouse_sensitivity() const { return mouse_sensitivity_; }
+  float cos_max_pitch_angle() const { return cos_max_pitch_angle_; }
 
-  void set_speed_per_sec(double value) { speed_per_sec_ = value; }
-  void set_mouse_sensitivity(double value) { mouse_sensitivity_ = value; }
-  void set_cos_max_pitch_angle(double value) { cos_max_pitch_angle_ = value; }
+  void set_speed_per_sec(float value) { speed_per_sec_ = value; }
+  void set_mouse_sensitivity(float value) { mouse_sensitivity_ = value; }
+  void set_cos_max_pitch_angle(float value) { cos_max_pitch_angle_ = value; }
 
  protected:
   bool first_call_;
-  double speed_per_sec_, mouse_sensitivity_, cos_max_pitch_angle_;
+  float speed_per_sec_, mouse_sensitivity_, cos_max_pitch_angle_;
 
  private:
   virtual void Update() override;
@@ -103,16 +103,16 @@ class FreeFlyCamera : public Camera {
 class ThirdPersonalCamera : public Camera {
  public:
   ThirdPersonalCamera(GameObject* parent,
-                      double fov,
-                      double z_near,
-                      double z_far,
-                      const glm::dvec3& position,
-                      double mouse_sensitivity = 1.0,
-                      double mouse_scroll_sensitivity = 1.0,
-                      double min_dist_mod = 0.25,
-                      double max_dist_mod = 4.00,
-                      double base_distance = 0.0,
-                      double dist_offset = 0.0);
+                      float fov,
+                      float z_near,
+                      float z_far,
+                      const glm::vec3& position,
+                      float mouse_sensitivity = 1.0,
+                      float mouse_scroll_sensitivity = 1.0,
+                      float min_dist_mod = 0.25,
+                      float max_dist_mod = 4.00,
+                      float base_distance = 0.0,
+                      float dist_offset = 0.0);
 
   virtual ~ThirdPersonalCamera() {}
 
@@ -124,12 +124,12 @@ class ThirdPersonalCamera : public Camera {
   bool first_call_;
 
   // Private constant numbers
-  const double initial_distance_, base_distance_, cos_max_pitch_angle_,
+  const float initial_distance_, base_distance_, cos_max_pitch_angle_,
                mouse_sensitivity_, mouse_scroll_sensitivity_,
                min_dist_mod_, max_dist_mod_, dist_offset_;
 
   // For mouseScrolled interpolation
-  double curr_dist_mod_, dest_dist_mod_;
+  float curr_dist_mod_, dest_dist_mod_;
 
   virtual void Update() override;
   virtual void MouseScrolled(double, double yoffset) override;
